fix(ollama): break sanitizer import cycle
