+++ extensions/ollama/src/sanitizers/visible-content-contract.ts
@@ -1,0 +1,9 @@
+export type OllamaVisibleContentStreamResolution =
+  | { kind: "visible"; text: string }
+  | { kind: "pending" };
+
+export type OllamaVisibleContentSanitizer = {
+  resolveStreamText(params: { text: string; final: boolean }): OllamaVisibleContentStreamResolution;
+  sanitizeFinalText(text: string): string;
+  shouldSanitizeFinalMessage(): boolean;
+};
